move core API functions out of view
